Add force groups parameter to API call.
This is exposed as a "force" prefix variable.
Ex: ?force.bgcolortst=2

forceGroups is necessary to include in the API so that web clients
can easily get payload values if they want to force a group. If this
wasn't in this API call, the client would have to dive into test
definition API calls just to handle the forceGroups special case.