Undo payload simplification.
In the /groups/identify API call, I previously made the payload
field return exactly whatever the payload field was, regardless
of the type. Now this is masked behind an object that maps, for
example, "stringArray" to ["one", "two"]. This is consistent
with how the rest of Proctor does it, and it might made it easier
for non-duck-typed languages to process responses.