Parse all force groups cookies
The previous behavior was to take the first prforceGroups cookie and use
that for the forced groups. If multiple different cookies are set across
multiple domains, this could result in groups being dropped along the
way. The change here is to set the forced groups as the maximal set of
all prforceGroups sent with this request.